Making complete model, added type declarations

+++ activations.jl
@@ -1,14 +1,14 @@
-function sigmoid(x)
+function sigmoid(x::Float64)::Float64
     return 1 / (1 + exp(-x))
 end
 
-function sigmoid_back(x)
+function sigmoid_back(x::Float64)::Float64
     sig = sigmoid(x)
     return sig*(1-sig)
 end
 
 
-function relu(x)
+function relu(x::Float64)::Float64
     if x>0
         return x
     else
@@ -16,7 +16,7 @@ function relu(x)
     end
 end
 
-function relu_back(x)
+function relu_back(x::Float64)::Float64
     if x>0
         return 1
     else
